supervisor/requirements.txt
requests==2.31.0
pyzmq==25.1.2
lxml==5.1.0